move test msg to cfg-traits
